Split GraphView.cs into sub files as suggested
