+++ Terminal.Gui/Core/Graphs/Orientation.cs
@@ -1,0 +1,17 @@
+﻿namespace Terminal.Gui.Graphs {
+	/// <summary>
+	/// Direction of an element (horizontal or vertical)
+	/// </summary>
+	public enum Orientation {
+
+		/// <summary>
+		/// Left to right 
+		/// </summary>
+		Horizontal,
+
+		/// <summary>
+		/// Bottom to top
+		/// </summary>
+		Vertical
+	}
+}
